requirements.txt
python-fasthtml @ git+https://github.com/AnswerDotAI/fasthtml.git
fh-matplotlib
fh-altair
great_tables
pandas
polars
fastsql
plotly
fh-plotly
numpy
fastcore
requests
seaborn
fh_frankenui @ git+https://github.com/AnswerDotAI/fh-frankenui.git
markdown2
lxml